Refactor: Android client toward API-driven architecture
- Replaced Room-based vehicle retrieval with Retrofit API integration
- Implemented API-backed vehicle deletion workflow
- Integrated backend authentication with JWT token persistence
- Introduced ApiClient and ApiService for centralized network handling
- Began aligning Vehicle model with backend schema (migrating ID from int to Long)
- Identified and addressed data layer inconsistencies between Room and API

+++ app/src/main/java/com/dylanclarke/FleetManagementApp/data/VehicleRepository.java
@@ -57,7 +57,7 @@ public class VehicleRepository {
 
     // Delete vehicle
     public boolean deleteVehicle(Vehicle vehicle) {
-        int count = db.maintenanceDao().countMaintenanceForVehicle(vehicle.getId());
+        int count = db.maintenanceDao().countMaintenanceForVehicle(vehicle.getId().intValue());
         if (count > 0) {
             return false;
         }
